Tidying

+++ src/cpu80186/iop80186.c
@@ -77,7 +77,7 @@ uint8_t portin(uint16_t portnum)
 
 uint16_t portin16(uint16_t portnum)
 {
-	return (uint16_t)( portin(portnum) | (portin(portnum + 1) << 8));
+	return (uint16_t)( portin(portnum) | (uint16_t) (portin(portnum + 1) << 8));
 }
 
 void portout16(uint16_t portnum, uint16_t value)
